:seedling: Update crypto structure
